refactor: SWRConfig api endPoint 수정 및 defaultValue 수정
